Use number of cycles for animation duration
All of the animations were expressed in number of cycles anyways, just using time. This saves a few bytes and is a little more efficient. Only downside is that an animation will need to complete (run all frames) before it will progress to the next linked animation, rather than expiring after the time period.